[REFACTOR-#9] - 도메인 역할 재조정을 위한 비즈니스 로직 변경으로 인해서 테스트 코드 전체 수정 - VerificationService Interface를 사용자 인증 정보 검증을 수행하는 명확한 역할로 할당   - 현재는 이메일 한정이라 메서드 명과 파라미터만 변경   - JavaDocs 추가
